# diag/openapi.json
{"openapi":"3.1.0","info":{"title":"Nutritional Assessment API","description":"API para evaluación y seguimiento nutricional infantil","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health Check","description":"Verifica el estado general y la conexión con la base de datos.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/healthz":{"get":{"summary":"Healthz Check","description":"Alias para Docker healthcheck","operationId":"healthz_check_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","description":"Endpoint raíz informativo","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/children/ping":{"get":{"tags":["children","children"],"summary":"Ping","operationId":"ping_api_children_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/children/":{"get":{"tags":["children","children"],"summary":"List Children","operationId":"list_children_api_children__get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChildOut"},"title":"Response List Children Api Children  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["children","children"],"summary":"Create Child","operationId":"create_child_api_children__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChildCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChildOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/children/{child_id}":{"get":{"tags":["children","children"],"summary":"Get Child","operationId":"get_child_api_children__child_id__get","parameters":[{"name":"child_id","in":"path","required":true,"schema":{"type":"integer","title":"Child Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChildOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["children","children"],"summary":"Update Child","operationId":"update_child_api_children__child_id__put","parameters":[{"name":"child_id","in":"path","required":true,"schema":{"type":"integer","title":"Child Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChildUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChildOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["children","children"],"summary":"Delete Child","operationId":"delete_child_api_children__child_id__delete","parameters":[{"name":"child_id","in":"path","required":true,"schema":{"type":"integer","title":"Child Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/register":{"post":{"tags":["auth","auth"],"summary":"Register","description":"Registra un nuevo usuario.","operationId":"register_api_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"tags":["auth","auth"],"summary":"Login","description":"Login con JSON:\n{\n  \"correo\": \"admin@example.com\",\n  \"contrasena\": \"12345678\"\n}","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/token":{"post":{"tags":["auth","auth"],"summary":"Token","description":"Login estilo OAuth2 con form-data:\n  username = correo\n  password = contrasena","operationId":"token_api_auth_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_token_api_auth_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/me":{"get":{"tags":["auth","auth"],"summary":"Me","operationId":"me_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/followups/ping":{"get":{"tags":["followups","followups"],"summary":"Ping","operationId":"ping_api_followups_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/followups/":{"get":{"tags":["followups","followups"],"summary":"List Followups","operationId":"list_followups_api_followups__get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FollowupOut"},"title":"Response List Followups Api Followups  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["followups","followups"],"summary":"Create Followup","operationId":"create_followup_api_followups__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FollowupCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FollowupOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/followups/{followup_id}":{"get":{"tags":["followups","followups"],"summary":"Get Followup","operationId":"get_followup_api_followups__followup_id__get","parameters":[{"name":"followup_id","in":"path","required":true,"schema":{"type":"integer","title":"Followup Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FollowupOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["followups","followups"],"summary":"Update Followup","operationId":"update_followup_api_followups__followup_id__put","parameters":[{"name":"followup_id","in":"path","required":true,"schema":{"type":"integer","title":"Followup Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FollowupUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FollowupOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["followups","followups"],"summary":"Delete Followup","operationId":"delete_followup_api_followups__followup_id__delete","parameters":[{"name":"followup_id","in":"path","required":true,"schema":{"type":"integer","title":"Followup Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/followups/{followup_id}/symptoms":{"post":{"tags":["followups","followups"],"summary":"Add Symptom","operationId":"add_symptom_api_followups__followup_id__symptoms_post","parameters":[{"name":"followup_id","in":"path","required":true,"schema":{"type":"integer","title":"Followup Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SymptomCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SymptomOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/followups/symptoms/{symptom_id}":{"get":{"tags":["followups","followups"],"summary":"Get Symptom","operationId":"get_symptom_api_followups_symptoms__symptom_id__get","parameters":[{"name":"symptom_id","in":"path","required":true,"schema":{"type":"integer","title":"Symptom Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SymptomOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reports/reports/":{"get":{"tags":["reports","Reports"],"summary":"Listado de endpoints de reportes","operationId":"reports_index_api_reports_reports__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/reports/reports/import/{import_id}/summary":{"get":{"tags":["reports","Reports"],"summary":"Resumen del dataset importado (conteos, promedios, BMI)","operationId":"import_summary_api_reports_reports_import__import_id__summary_get","parameters":[{"name":"import_id","in":"path","required":true,"schema":{"type":"string","description":"ID devuelto por /import/excel","title":"Import Id"},"description":"ID devuelto por /import/excel"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reports/reports/import/{import_id}/export":{"get":{"tags":["reports","Reports"],"summary":"Exportar Excel con cálculos (edad, BMI)","operationId":"export_enriched_excel_api_reports_reports_import__import_id__export_get","parameters":[{"name":"import_id","in":"path","required":true,"schema":{"type":"string","description":"ID devuelto por /import/excel","title":"Import Id"},"description":"ID devuelto por /import/excel"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/import/excel":{"post":{"tags":["import","Importación Excel"],"summary":"Importar Excel (vectoriza y persiste por import_id)","operationId":"import_excel_file_api_import_excel_post","parameters":[{"name":"text_columns","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"CSV de columnas a usar como texto. Si omites, concatena todas.","title":"Text Columns"},"description":"CSV de columnas a usar como texto. Si omites, concatena todas."}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_import_excel_file_api_import_excel_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/import/template":{"get":{"tags":["import","Importación Excel"],"summary":"Descargar plantilla Excel","operationId":"download_template_api_import_template_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/import/status/{import_id}":{"get":{"tags":["import","Importación Excel"],"summary":"Consultar estado de import","operationId":"get_import_status_api_import_status__import_id__get","parameters":[{"name":"import_id","in":"path","required":true,"schema":{"type":"string","title":"Import Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/import/search":{"get":{"tags":["import","Importación Excel"],"summary":"Buscar por similitud en un import_id","operationId":"search_in_import_api_import_search_get","parameters":[{"name":"import_id","in":"query","required":true,"schema":{"type":"string","description":"ID devuelto por /import/excel","title":"Import Id"},"description":"ID devuelto por /import/excel"},{"name":"q","in":"query","required":true,"schema":{"type":"string","description":"Consulta de texto","title":"Q"},"description":"Consulta de texto"},{"name":"top_k","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"default":5,"title":"Top K"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SearchResult"},"title":"Response Search In Import Api Import Search Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_import_excel_file_api_import_excel_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_import_excel_file_api_import_excel_post"},"Body_token_api_auth_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_token_api_auth_token_post"},"ChildCreate":{"properties":{"nombre":{"type":"string","minLength":1,"title":"Nombre","examples":["Ana María Pérez Rojas"]},"fecha_nacimiento":{"type":"string","pattern":"^\\d{4}-\\d{2}-\\d{2}$","title":"Fecha Nacimiento","examples":["2021-05-10"]},"genero":{"type":"string","minLength":1,"title":"Genero","examples":["F","M"]},"sede_id":{"type":"integer","minimum":1.0,"title":"Sede Id","examples":[1]},"acudiente_id":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Acudiente Id"}},"type":"object","required":["nombre","fecha_nacimiento","genero","sede_id"],"title":"ChildCreate"},"ChildOut":{"properties":{"id":{"type":"integer","title":"Id"},"nombre":{"type":"string","title":"Nombre"},"fecha_nacimiento":{"type":"string","title":"Fecha Nacimiento"},"genero":{"type":"string","title":"Genero"},"sede_id":{"type":"integer","title":"Sede Id"},"acudiente_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Acudiente Id"}},"type":"object","required":["id","nombre","fecha_nacimiento","genero","sede_id"],"title":"ChildOut"},"ChildUpdate":{"properties":{"nombre":{"anyOf":[{"type":"string","minLength":1},{"type":"null"}],"title":"Nombre"},"fecha_nacimiento":{"anyOf":[{"type":"string","pattern":"^\\d{4}-\\d{2}-\\d{2}$"},{"type":"null"}],"title":"Fecha Nacimiento"},"genero":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Genero"},"sede_id":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Sede Id"},"acudiente_id":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Acudiente Id"}},"type":"object","title":"ChildUpdate"},"FollowupCreate":{"properties":{"child_id":{"type":"integer","minimum":1.0,"title":"Child Id"},"fecha":{"type":"string","pattern":"^\\d{4}-\\d{2}-\\d{2}$","title":"Fecha","examples":["2024-09-01"]},"peso_kg":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Peso Kg"},"talla_cm":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Talla Cm"},"observaciones":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Observaciones"}},"type":"object","required":["child_id","fecha"],"title":"FollowupCreate"},"FollowupOut":{"properties":{"id":{"type":"integer","title":"Id"},"child_id":{"type":"integer","title":"Child Id"},"fecha":{"type":"string","title":"Fecha"},"peso_kg":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Peso Kg"},"talla_cm":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Talla Cm"},"observaciones":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Observaciones"}},"type":"object","required":["id","child_id","fecha"],"title":"FollowupOut"},"FollowupUpdate":{"properties":{"fecha":{"anyOf":[{"type":"string","pattern":"^\\d{4}-\\d{2}-\\d{2}$"},{"type":"null"}],"title":"Fecha"},"peso_kg":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Peso Kg"},"talla_cm":{"anyOf":[{"type":"number","minimum":0.0},{"type":"null"}],"title":"Talla Cm"},"observaciones":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Observaciones"}},"type":"object","title":"FollowupUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"correo":{"type":"string","format":"email","title":"Correo"},"contrasena":{"type":"string","title":"Contrasena"}},"type":"object","required":["correo","contrasena"],"title":"LoginRequest"},"MeResponse":{"properties":{"id_usuario":{"type":"integer","title":"Id Usuario"},"nombre":{"type":"string","title":"Nombre"},"correo":{"type":"string","format":"email","title":"Correo"},"telefono":{"type":"string","title":"Telefono"}},"type":"object","required":["id_usuario","nombre","correo","telefono"],"title":"MeResponse"},"SearchResult":{"properties":{"row_index":{"type":"integer","title":"Row Index"},"score":{"type":"number","title":"Score"},"payload":{"additionalProperties":{"type":"string"},"type":"object","title":"Payload"}},"type":"object","required":["row_index","score","payload"],"title":"SearchResult"},"SymptomCreate":{"properties":{"codigo":{"type":"string","minLength":1,"title":"Codigo","examples":["fiebre"]},"severidad":{"anyOf":[{"type":"integer","maximum":5.0,"minimum":1.0},{"type":"null"}],"title":"Severidad"}},"type":"object","required":["codigo"],"title":"SymptomCreate"},"SymptomOut":{"properties":{"id_symptom":{"type":"integer","title":"Id Symptom"},"followup_id":{"type":"integer","title":"Followup Id"},"codigo":{"type":"string","title":"Codigo"},"severidad":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Severidad"}},"type":"object","required":["id_symptom","followup_id","codigo"],"title":"SymptomOut"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"expires_in":{"type":"integer","title":"Expires In"}},"type":"object","required":["access_token","expires_in"],"title":"TokenResponse"},"UsuarioCreate":{"properties":{"nombre":{"type":"string","title":"Nombre"},"correo":{"type":"string","format":"email","title":"Correo"},"telefono":{"type":"string","title":"Telefono"},"contrasena":{"type":"string","title":"Contrasena"}},"type":"object","required":["nombre","correo","telefono","contrasena"],"title":"UsuarioCreate"},"UsuarioResponse":{"properties":{"nombre":{"type":"string","title":"Nombre"},"correo":{"type":"string","format":"email","title":"Correo"},"telefono":{"type":"string","title":"Telefono"},"id_usuario":{"type":"integer","title":"Id Usuario"},"fecha_creado":{"type":"string","format":"date-time","title":"Fecha Creado"}},"type":"object","required":["nombre","correo","telefono","id_usuario","fecha_creado"],"title":"UsuarioResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/auth/token"}}}}}}